requirements.txt
jupyter
selenium
wget
opencv-python
numpy
sklearn
imutils
keras 
tensorflow 
pillow
matplotlib
